requirements.txt
pydantic==2.2.1
starlette==0.27.0   
redis==5.0.0
celery==5.3.1        
selenium==4.11.2
kafka-python==2.0.2
python_dotenv==1.0.0
typing-inspect==0.9.0
pydantic-settings==2.0.3
urllib3==1.26.7
pytube==15.0.0
dateparser
requests
jsonpath_ng
elasticsearch
colorlog
schedule
ctypes==1.1.0
python-dotenv==1.0.0